models CTA click report as a thing not just mooshed over various repositories
